Handle Clang's more picky interpretation of GCC atomic builtins.
The "picky" behavior was introduced in trunk before the 3.0 branch,
so we test for a major version of 3 or greater to enable to workaround.

cf. http://llvm.org/bugs/show_bug.cgi?id=11280

Should be safe since a const_cast typically does not emit code, but
enabled just for the pickier clang versions to be cautious.